Additional renames to fix compilation errors
