+++ src/main/java/seedu/findvisor/commons/util/DateTimeUtil.java
@@ -1,4 +1,4 @@
-package seedu.address.commons.util;
+package seedu.findvisor.commons.util;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
