+++ src/main/java/seedu/findvisor/logic/commands/ScheduleCommand.java
@@ -1,27 +1,27 @@
-package seedu.address.logic.commands;
+package seedu.findvisor.logic.commands;
 
 import static java.util.Objects.requireNonNull;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_END_DATETIME;
-import static seedu.address.logic.parser.CliSyntax.PREFIX_START_DATETIME;
-import static seedu.address.model.Model.PREDICATE_SHOW_ALL_PERSONS;
+import static seedu.findvisor.logic.parser.CliSyntax.PREFIX_END_DATETIME;
+import static seedu.findvisor.logic.parser.CliSyntax.PREFIX_START_DATETIME;
+import static seedu.findvisor.model.Model.PREDICATE_SHOW_ALL_PERSONS;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
-import seedu.address.commons.core.index.Index;
-import seedu.address.commons.util.DateTimeUtil;
-import seedu.address.commons.util.ToStringBuilder;
-import seedu.address.logic.Messages;
-import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.Model;
-import seedu.address.model.person.Address;
-import seedu.address.model.person.Email;
-import seedu.address.model.person.Meeting;
-import seedu.address.model.person.Name;
-import seedu.address.model.person.Person;
-import seedu.address.model.person.Phone;
-import seedu.address.model.tag.Tag;
+import seedu.findvisor.commons.core.index.Index;
+import seedu.findvisor.commons.util.DateTimeUtil;
+import seedu.findvisor.commons.util.ToStringBuilder;
+import seedu.findvisor.logic.Messages;
+import seedu.findvisor.logic.commands.exceptions.CommandException;
+import seedu.findvisor.model.Model;
+import seedu.findvisor.model.person.Address;
+import seedu.findvisor.model.person.Email;
+import seedu.findvisor.model.person.Meeting;
+import seedu.findvisor.model.person.Name;
+import seedu.findvisor.model.person.Person;
+import seedu.findvisor.model.person.Phone;
+import seedu.findvisor.model.tag.Tag;
 
 /**
  * Schedules a meeting with a contact.
